style: 구현방법의 간단한 주석설명 추가, ItemDecoration 파일 Rename
