+++ app/src/main/java/com/example/andromedaproject/ui/MainActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.andromedaproject.R
-import com.example.andromedaproject.Utils.ItemDecoration
+import com.example.andromedaproject.Utils.ItemDivider
 import com.example.andromedaproject.adapter.DragManageAdapter
 import com.example.andromedaproject.adapter.ProfileAdapter
 import com.example.andromedaproject.adapter.SwipeToDeleteAdapter
@@ -29,7 +29,14 @@ class MainActivity : AppCompatActivity() {
         profileAdapter = ProfileAdapter(this)
         recyclerview_profile.adapter = profileAdapter
         recyclerview_profile.layoutManager = LinearLayoutManager(this)
-        recyclerview_profile.addItemDecoration(ItemDecoration(5, 5))
+
+        //리사이클러뷰의 아이템들을 높이 5dp, 좌우로 5dp를 띄우겠다는 뜻입니다.
+        //Utils 파일에 ItemDivider를 만들어 놓았으니 참고해주세요!!
+        //리사이클러뷰 이름에 addItemDecoration이라는 함수를 통해서 설정해줍니다.
+        //addItemDecoration을 ctrl클릭해보면 생성자로 ItemDecoration을 넣어주여야하는데
+        //우리는 아까 ItemDivider를 만들 때 ItemDecoration을 상속시켰으니
+        //addItemDecoration() 여기 괄호안에 ItemDivider를 넣어주면 됩니다.
+        recyclerview_profile.addItemDecoration(ItemDivider(5, 5))
         moveItemHelper()
         deleteItemHelper()
         loadDatas()
